+++ phy/cluster/__init__.py
@@ -2,12 +2,3 @@
 # flake8: noqa
 
 """Automatic and manual clustering facilities."""
-
-from .view_models import (BaseClusterViewModel,
-                          HTMLClusterViewModel,
-                          StatsViewModel,
-                          FeatureViewModel,
-                          WaveformViewModel,
-                          TraceViewModel,
-                          CorrelogramViewModel,
-                          )
