+++ phy/cluster/manual/__init__.py
@@ -1,10 +1,16 @@
 # -*- coding: utf-8 -*-
 # flake8: noqa
 
-"""Manual clustering facilities.
+"""Manual clustering facilities."""
 
-"""
-
+from .view_models import (BaseClusterViewModel,
+                          HTMLClusterViewModel,
+                          StatsViewModel,
+                          FeatureViewModel,
+                          WaveformViewModel,
+                          TraceViewModel,
+                          CorrelogramViewModel,
+                          )
 from .clustering import Clustering
 from .wizard import Wizard
 from .gui import ClusterManualGUI
