+++ phy/cluster/manual/gui.py
@@ -13,13 +13,13 @@ import numpy as np
 import phy
 from ...gui.base import BaseGUI
 from ...gui.qt import _prompt
-from ..view_models import (WaveformViewModel,
-                           FeatureGridViewModel,
-                           FeatureViewModel,
-                           CorrelogramViewModel,
-                           TraceViewModel,
-                           StatsViewModel,
-                           )
+from .view_models import (WaveformViewModel,
+                          FeatureGridViewModel,
+                          FeatureViewModel,
+                          CorrelogramViewModel,
+                          TraceViewModel,
+                          StatsViewModel,
+                          )
 from ...utils.logging import debug, info, warn
 from ...io.kwik.model import cluster_group_id
 from ._history import GlobalHistory
