+++ phy/session/session.py
@@ -20,7 +20,7 @@ from ..io.kwik.model import KwikModel
 from ..io.kwik.store_items import create_store
 from ..cluster.manual.gui import ClusterManualGUI
 from ..cluster.algorithms.klustakwik import KlustaKwik
-from ..detect import SpikeDetekt
+from ..detect.spikedetekt import SpikeDetekt
 
 
 #------------------------------------------------------------------------------
